+++ src/tightdb/objc/test/table_view.m
@@ -98,16 +98,16 @@
     TDBTable *t = [[TDBTable alloc] init];
     NSUInteger intCol = [t addColumnWithName:@"intCol" type:TDBIntType];
     
-    NSUInteger rowIndex = [t addRow:nil];
-    TDBRow *row = [t rowAtIndex:rowIndex];
+    [t addRow:nil];
+    TDBRow *row = [t lastRow];
     [row setInt:2 inColumnWithIndex:intCol];
     
-    rowIndex = [t addRow:nil];
-    row = [t rowAtIndex:rowIndex];
+    [t addRow:nil];
+    row = [t lastRow];
     [row setInt:1 inColumnWithIndex:intCol];
     
-    rowIndex = [t addRow:nil];
-    row = [t rowAtIndex:rowIndex];
+    [t addRow:nil];
+    row = [t lastRow];
     [row setInt:0 inColumnWithIndex:intCol];
     
     TDBQuery *q = [t where];
@@ -141,14 +141,17 @@
 {
     TDBTable *t = [[TDBTable alloc] init];
     NSUInteger boolCol = [t addColumnWithName:@"boolCol" type:TDBBoolType];
-    
-    TDBRow *row = [t rowAtIndex:[t addRow:nil]];
+
+    [t addRow:nil];
+    TDBRow *row = [t lastRow];
     [row setBool:YES inColumnWithIndex:boolCol];
-    
-    row = [t rowAtIndex:[t addRow:nil]];
+
+    [t addRow:nil];
+    row = [t lastRow];
     [row setBool:YES inColumnWithIndex:boolCol];
-    
-    row = [t rowAtIndex:[t addRow:nil]];
+
+    [t addRow:nil];
+    row = [t lastRow];
     [row setBool:NO inColumnWithIndex:boolCol];
     
     TDBQuery *q = [t where];
@@ -192,16 +195,16 @@
     NSDate *dateMiddle  = [formatter dateFromString:@"02/01/2014 10:10 PM"];
     NSDate *dateLast    = [formatter dateFromString:@"03/01/2014 10:10 PM"];
     
-    NSUInteger rowIndex = [t addRow:nil];
-    TDBRow *row = [t rowAtIndex:rowIndex];
+    [t addRow:nil];
+    TDBRow *row = [t lastRow];
     [row setDate:dateLast inColumnWithIndex:dateCol];
     
-    rowIndex = [t addRow:nil];
-    row = [t rowAtIndex:rowIndex];
+    [t addRow:nil];
+    row = [t lastRow];
     [row setDate:dateMiddle inColumnWithIndex:dateCol];
     
-    rowIndex = [t addRow:nil];
-    row = [t rowAtIndex:rowIndex];
+    [t addRow:nil];
+    row = [t lastRow];
     [row setDate:dateFirst inColumnWithIndex:dateCol];
     
     TDBQuery *q = [t where];
@@ -312,6 +315,33 @@
     XCTAssertTrue([view[0][0] isEqual:@10], @"row 0 -> 0");
     XCTAssertTrue([view[1][0] isEqual:@27], @"row 1 -> 2");
     XCTAssertTrue([view[2][0] isEqual:@8],  @"row 2 -> 4");
+}
+
+- (void)testQueryOnView
+{
+    TDBTable *table = [[TDBTable alloc] init];
+    
+    // Specify the column types and names
+    [table addColumnWithName:@"firstName" type:TDBStringType];
+    [table addColumnWithName:@"lastName" type:TDBStringType];
+    [table addColumnWithName:@"salary" type:TDBIntType];
+    
+    // Add data to the table
+    [table addRow:@[@"John", @"Lee", @10000]];
+    [table addRow:@[@"Jane", @"Lee", @15000]];
+    [table addRow:@[@"John", @"Anderson", @20000]];
+    [table addRow:@[@"Erik", @"Lee", @30000]];
+    [table addRow:@[@"Henry", @"Anderson", @10000]];
+    
+    
+    TDBView *view = [[table where] findAllRows];
+    XCTAssertEqual(view.rowCount, (NSUInteger)5, @"All 5 rows still here");
+
+    TDBView *view2 = [[[view where ] stringIsCaseInsensitiveEqualTo:@"John" inColumnWithIndex:0 ] findAllRows];
+    XCTAssertEqual(view2.rowCount, (NSUInteger)2, @"2 rows match");
+    
+    TDBView *view3 = [[[view2 where] stringIsCaseInsensitiveEqualTo:@"Anderson" inColumnWithIndex:1 ] findAllRows];
+    XCTAssertEqual(view3.rowCount, (NSUInteger)1, @"Only 1 row left");
 }
 
 @end